Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main () {
-    int n ;
-    scanf("%d",&n);
+    float n ;
+    scanf("%f",&n);
     float area = n * n * 3.14 ;
     printf("Area: %.2f",area);
 }
